+++ src/main/kotlin/im/mrx/leolanguage/leo/reference/LeoExternalRecordReference.kt
@@ -20,11 +20,11 @@ import com.intellij.openapi.project.DumbService
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiReference
 import com.intellij.psi.impl.source.resolve.ResolveCache
-import im.mrx.leolanguage.leo.psi.LeoExternalFunctionIdentifier
+import im.mrx.leolanguage.leo.psi.LeoExternalRecord
 import im.mrx.leolanguage.leo.stub.IndexUtils
 
-class LeoExternalFunctionReference(element: LeoExternalFunctionIdentifier) :
-    LeoReferenceBase<LeoExternalFunctionIdentifier>(element) {
+class LeoExternalRecordReference(element: LeoExternalRecord) :
+    LeoReferenceBase<LeoExternalRecord>(element) {
 
     override fun resolve(): PsiElement? {
         return ResolveCache.getInstance(element.project).resolveWithCaching(this, Resolver, false, false)
@@ -35,7 +35,7 @@ class LeoExternalFunctionReference(element: LeoExternalFunctionIdentifier) :
 
         override fun resolve(ref: PsiReference, incompleteCode: Boolean): PsiElement? {
             if (DumbService.isDumb(ref.element.project)) return null
-            val element = ref.element as LeoExternalFunctionIdentifier
+            val element = ref.element as LeoExternalRecord
             val file =
                 element.containingFile.containingDirectory.parentDirectory?.findSubdirectory("imports")
                     ?.findFile(element.programId.text) ?: return null